+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Changes
 -  Run GIF downloader asynchronously so users won't get stuck
 -  Reply to user message on GIF download, so that the user sees to which GIF the message belongs
 -  Improve TTS error message
+-  Rename ``tty`` command to ``tts`` (Text-To-Speech) but add an ``tty`` alias for the time being
 
 
 [1.4.0] - 2018.05.18